Refactors EvernoteNote/EvernoteRequest and Note/Citation.

+++ app/models/note.rb
@@ -2,9 +2,11 @@
 
 class Note < ActiveRecord::Base
 
+  include SyncHelper
+
   attr_accessible :title, :body, :external_updated_at, :resources, :latitude, :longitude, :lang, :author,
                   :last_edited_by, :source, :source_application, :source_url, :sources, :tag_list, :instruction_list,
-                  :hide, :active
+                  :hide, :active, :is_citation, :listable
 
   attr_writer :tag_list, :instruction_list
 
@@ -15,30 +17,32 @@ class Note < ActiveRecord::Base
   acts_as_taggable_on :tags, :instructions
 
   has_paper_trail on: [:update],
+                  unless: proc { |note| note.has_instruction?('reset') },
                   meta: {
                     sequence:  proc { |note| note.versions.length + 1 },  # To retrieve by version number
                     tag_list:  proc { |note| Note.find(note.id).tag_list }, # Note.tag_list would store incoming tags
                     instruction_list:  proc { |note| Note.find(note.id).instruction_list }
                   }
 
-  scope :publishable, where('active = ? AND hide = ?', true, false)
   default_scope order: 'external_updated_at DESC'
+  scope :publishable, where(active: true, hide: false)
+  scope :listable, where(listable: true, is_citation: false)
+  scope :citations, where(is_citation: true)
 
   validates :title, :external_updated_at, presence: true
   validate :body_or_source_or_resource, before: :update
+  # Being activated on create and throws error
+  # validate :external_updated_at_must_be_latest, :before => :update
 
-  # FIXME: Doesn't work with :body_changed?
-  before_validation :scan_note_for_references # , :if => :body_changed?
-  before_validation :scan_note_for_isbns      # , :if => :body_changed?
+  before_save :discard_versions? #, :if => :instruction_list_changed?
+  before_save :scan_note_for_references, :if => :body_changed?
+  after_save :scan_note_for_isbns, :if => :body_changed?
 
   def body_or_source_or_resource
     if body.blank? && embeddable_source_url.blank? && resources.blank?
       errors.add(:note, 'Note needs one of body, source or resource.')
     end
   end
-
-  # Being activated on create and throws error
-  # validate :external_updated_at_must_be_latest, :before => :update
 
   def headline
     I18n.t('notes.untitled_synonyms').include?(title) ? I18n.t('notes.short', id: id) : title
@@ -53,6 +57,12 @@ class Note < ActiveRecord::Base
     end
   end
 
+  def citation_blurb
+    citation_blurb = body.gsub(/^.*quote:/, '')
+    attribution = citation_blurb.scan(/--.*?$/m).first
+    "#{ citation_blurb[0 .. Settings.notes.blurb_length] }<br>#{ attribution }"
+  end
+
   def embeddable_source_url
     if source_url && source_url =~ /youtube|vimeo|soundcloud/
       source_url
@@ -64,16 +74,10 @@ class Note < ActiveRecord::Base
     end
   end
 
-  def instructed_to?(instruction, instructions = instruction_list)
-    (Settings.notes.instructions + instructions).include? "__#{ instruction.upcase }"
-  end
-
   def fx
-    fx = (Settings.notes.instructions + Array(instruction_list)).keep_if { |i| i =~ /__FX_/ }
-                                                         .join('_')
-                                                         .gsub(/__FX_/, '')
-                                                         .downcase
-    fx == '' ? nil : fx
+    instructions = Settings.notes.instructions.default + Array(instruction_list)
+    fx = instructions.keep_if { |i| i =~ /__FX_/ } .join('_').gsub(/__FX_/, '').downcase
+    fx.empty? ? nil : fx
   end
 
   # private
@@ -84,52 +88,17 @@ class Note < ActiveRecord::Base
     #     end
     # end
 
-  def update_with_evernote_data(note_data, cloud_note_tags)
-    update_attributes!(
-      title: note_data.title,
-      body: note_data.content,
-      lang: lang_from_cloud("#{ note_data.title } #{ note_data.content }"),
-      latitude: note_data.attributes.latitude,
-      longitude: note_data.attributes.longitude,
-      external_updated_at: calculate_updated_at(note_data, cloud_note_tags),
-      author: note_data.attributes.author,
-      last_edited_by: note_data.attributes.lastEditedBy,
-      source: note_data.attributes.source,
-      source_application: note_data.attributes.sourceApplication,
-      source_url: note_data.attributes.sourceURL,
-      tag_list: cloud_note_tags.grep(/^[^_]/),
-      instruction_list: cloud_note_tags.grep(/^_/),
-      hide: instructed_to?('hide'),
-      active: true
-    )
-    update_resources_with_evernote_data(note_data)
-  end
-
-  def lang_from_cloud(content)
-    (ActionController::Base.helpers.strip_tags(content[0..Settings.notes.wtf_sample_length])).lang
-  end
-
-  def calculate_updated_at(note_data, cloud_note_tags)
-    use_date = note_data.updated
-    # REVIEW: this method does two things
-    # Also, a version is still created after saving.
-    unless (Settings.evernote.instructions.reset & cloud_note_tags).empty?
-      use_date = note_data.created
-      versions.destroy_all
-    end
-
-    use_date = note_data.created if external_updated_at.blank? && Settings.evernote.reset
-
-    Time.at(use_date / 1000).to_datetime
-  end
-
   def scan_note_for_references
     # REVIEW: Should this be in Book?
-    self.books = Book.citable.keep_if { |book| self.body.include?(book.tag) }
+    self.books = Book.citable.keep_if { |book| body.include?(book.tag) }
   end
 
   def scan_note_for_isbns
     Book.grab_isbns(body) unless body.blank?
+  end
+
+  def discard_versions?
+    versions.destroy_all if has_instruction?('reset')
   end
 
   # REVIEW: When a note contains both images and downloads, the alt/cap is disrupted
